convert Index, QuestIndex, ObjectiveIndex to uint32

+++ packages/contracts/src/libraries/LibRegistryQuests.sol
@@ -40,7 +40,7 @@ library LibRegistryQuests {
   function createQuest(
     IWorld world,
     IUintComp components,
-    uint256 index,
+    uint32 index,
     string memory name,
     string memory description
   ) internal returns (uint256) {
@@ -65,7 +65,7 @@ library LibRegistryQuests {
   function createEmptyObjective(
     IWorld world,
     IUintComp components,
-    uint256 questIndex,
+    uint32 questIndex,
     string memory name, // this is a crutch to help FE, ideally we drop this
     string memory logicType,
     string memory type_
@@ -79,14 +79,14 @@ library LibRegistryQuests {
     setType(components, id, type_);
 
     uint256 numObjectives = getObjectivesByQuestIndex(components, questIndex).length;
-    setObjectiveIndex(components, id, numObjectives + 1);
+    setObjectiveIndex(components, id, uint32(numObjectives) + 1);
     return id;
   }
 
   function createEmptyRequirement(
     IWorld world,
     IUintComp components,
-    uint256 questIndex,
+    uint32 questIndex,
     string memory logicType,
     string memory type_
   ) internal returns (uint256) {
@@ -102,7 +102,7 @@ library LibRegistryQuests {
   function createEmptyReward(
     IWorld world,
     IUintComp components,
-    uint256 questIndex,
+    uint32 questIndex,
     string memory type_
   ) internal returns (uint256) {
     uint256 id = world.getUniqueEntityId();
@@ -210,15 +210,15 @@ library LibRegistryQuests {
     IsRewardComponent(getAddressById(components, IsRewardCompID)).set(id);
   }
 
-  function setIndex(IUintComp components, uint256 id, uint256 index) internal {
+  function setIndex(IUintComp components, uint256 id, uint32 index) internal {
     IndexComponent(getAddressById(components, IndexCompID)).set(id, index);
   }
 
-  function setObjectiveIndex(IUintComp components, uint256 id, uint256 index) internal {
+  function setObjectiveIndex(IUintComp components, uint256 id, uint32 index) internal {
     IndexObjectiveComponent(getAddressById(components, IndexObjectiveCompID)).set(id, index);
   }
 
-  function setQuestIndex(IUintComp components, uint256 id, uint256 questIndex) internal {
+  function setQuestIndex(IUintComp components, uint256 id, uint32 questIndex) internal {
     IndexQuestComponent(getAddressById(components, IndexQuestCompID)).set(id, questIndex);
   }
 
@@ -355,7 +355,7 @@ library LibRegistryQuests {
   // get registry entry by Quest index
   function getByQuestIndex(
     IUintComp components,
-    uint256 index
+    uint32 index
   ) internal view returns (uint256 result) {
     QueryFragment[] memory fragments = new QueryFragment[](3);
     fragments[0] = QueryFragment(QueryType.Has, getComponentById(components, IsRegCompID), "");
@@ -373,7 +373,7 @@ library LibRegistryQuests {
   // get Objectives by Quest index
   function getObjectivesByQuestIndex(
     IUintComp components,
-    uint256 index
+    uint32 index
   ) internal view returns (uint256[] memory results) {
     QueryFragment[] memory fragments = new QueryFragment[](3);
     fragments[0] = QueryFragment(QueryType.Has, getComponentById(components, IsRegCompID), "");
@@ -394,7 +394,7 @@ library LibRegistryQuests {
   // get requirements by Quest index
   function getRequirementsByQuestIndex(
     IUintComp components,
-    uint256 index
+    uint32 index
   ) internal view returns (uint256[] memory results) {
     QueryFragment[] memory fragments = new QueryFragment[](3);
     fragments[0] = QueryFragment(QueryType.Has, getComponentById(components, IsRegCompID), "");
@@ -415,7 +415,7 @@ library LibRegistryQuests {
   // get reward by Quest index
   function getRewardsByQuestIndex(
     IUintComp components,
-    uint256 index
+    uint32 index
   ) internal view returns (uint256[] memory results) {
     QueryFragment[] memory fragments = new QueryFragment[](3);
     fragments[0] = QueryFragment(QueryType.Has, getComponentById(components, IsRegCompID), "");
